Click: Fixup the best road selection
The road was not computed again in case the player's attack decreased
This case is now handled properly.

Add the safary state to the instance states

Improve UseOakItemQuest loadout selection

Improve best route for a given pokemon type, it now chose the route
base on the given type pokemon rate instead of count.