Quest: Add an option to auto-restore mining energy
Mining quests can be pretty slow to complete with the regular energy
refill.

A button was added to allow buying and using small restore potions
automatically during mining quests.

The refill potion cannot be bought until the Cinnabar Island city
is unlocked, so the auto-buy is disabled if this city is not unlocked.