Dungeon: Fixup a dead-lock
If a module set the __stopRequested member and did not reset it
before being disabled, the dungeon would never be able to start
anymore.

The Quest module now resets the value while turned off.
The Dungeon module sets the value to false after taking the
demand into account to avoid blocking any further.