Menu: Add a reason when disabling buttons
The 'Quests' feature disables other mode while running.
The 'Got to' feature self disable when the player can't de moved.

The end-user might not understand why the button where disabled.
The tooltip now indicates why.

The auto-dungeon button is now disabled when the player can't enter
the dungeon for one of the following reason:
   - The Dungeon ticket was not bought
   - The pokedex-only option is on and all pokemons are already caught
   - The dungeon entrance fee can't be payed