Quest: Fixup UseOakItemQuest not using the right item
If the quest targeted an item that was part of the default loadout,
but it was at a position in the list that was past the player's
max concurrent item count, the item would never be equipped.

The item is now systematically removed from the default loadout, and
added back to the front of it.
This will guaranty that the item is indeed equipped.