Mine: Don't run multiple loops at the same time
If a loop takes more than 10s to consume all the payer's energy,
an additional loop is executed.

This could happen if the player or the quest focus topic refreshed
energy while the loop was running.

This would speed-up the mining process (and might mess it up), and
send multiple times the completion notification.

---

In addition, using a tool without having enough energy did not
stop the loop and did increment the usage counter.
This resulted in weird notifications with very high usage count.