Dungeon: Fixup player rushing to boss in some rare cases
If the player has the flashlight (after defeating the dungeon more
than 200 times), the tiles nearby the player are visible.
When entering a dungeon, if the boss is on a tile directly adjacent
to the entrance, the script would rush for it immediately, since it's
visible.

In case the player never moved, it's not the intended behaviour.
The dungeon is now only tagged as completed if either:
- It's not the first moved
- Or:
  - The boss tile is visible or visited
  - The player visited at least a tile (apart from the entrance)