Focus.Achievements: Fixup the achievements filtering method
The gym can be listed because the town was unlocked,
but still be locked.
The gym unlocked state is now checked in the achievements filter.

---

On Chrome, the initial order was not preserved after the reorder call.
This was due to not handling equals elements properly, which is the
only way to preserve the order.