Loader: Don't load and run before the game is fully instanciated
Running the game on a low-end environment (a raspberry pi for instance)
can take a significant amount of time before it's up and running.

This could cause the menu to request game components that are not yet
existing or instantiated.

The loader now makes sure the game is properly loaded before loading
the automation.