Farm: Fix the harvesting method while doing quests
While the `Focus on quest` feature is enabled, the script will switch
to Cheri berries to avoid occupying the farm with long riping ones.
It also asks the Farming automation to harvest berries as fast as
possible.

However, this parameter was not considered by the harvest method.
It's now properly done.