Hatchery: Add Pokérus speading strategy
This option will add contagious Pokémon if none are present in the
hatchery.
It will then add Pokérus-free ones matching the contagious pokémon type
to spread the pokerus to them.

Once every pokemon have been infected, it will fallback to the default
strategy.

This option is togglable in the advanced settings and is on by default.
Since the Pokérus is an end-game unlock, the option is disabled until
the player reached this point.