Utils.Route: Optimize best route for dungeon token method
The pokeclicker `calcBonus` method is performance-heavy.
The `moveToHighestDungeonTokenIncomeRoute` method spent half
the time in it.

The call to this method is used to retrieve each route's dungeon
token income.
Such value never changes across the game's life.
Those values are now computed for every routes upon instantiation
of the class and stored into a dedicated map.