Focus.Quest: Fixup UseSmallRestore not working
The restore did not happen when asked for.
This was due to an incorrect condition.

The condition has been fixed and the functionality moved to the
Underground class, since it's underground related.

A UseSmallRestoreAsked Setting was added to this class to allow
other classes (the Focus.Quest one for now) to ask for it.