Farm: Fixup oak item loadout being messed-up
The loadout was always reset because the item id was compared to
the item object, which was obviously always different.

The custom loadout was a mix of items and item id.
Thus, only the needed item was equipped back because the `activate()`
method takes the item's id as param, not the item object.

Only the last item of the current loadout is removed now, only if there
is no room for the needed item.