Utils.Route: Improve best dungeon token route computation
In order to gain dungeon token, pokémon needs to be captured.
The catch rate was not taken into account while determining the best
route to farm tokens. This could result in a less efficient road
selection, and thus a drop in token income.

The route average catch rate is now taken into account by the
algorithm.