Shop: Allow some items to be bought once unlocked
Initially the auto shop feature was only available once the Poké Mart
in-game shop was unlocked by the player.

The items are now unlocked as soon as they can be purchased from an
unlocked shop.

Note: The feature will not purchase items if they are not available in
any accessible shop (after moving to a new region for instance).