Hatchery: Add a type priority setting
The player can now choose to breed pokémons from a specific
type in priority.

If a region priority is set as well, and no pokémon of the selected
region and type can be hatched, the algorithm will select pokémons
from the such region before ones matching the type.