Farm: Wait for berries withering while trying to unlock Kasib
The Kasib berry requires any berry to wither to have a chance to
appear in the plot.
The regular strategy was used for this unlock, which plants Cheri
berries. However, those were harvested right before dying, instead
of letting those die.

The harvest strategy now contains the following possible appoach:
  - AsSoonAsPossible
  - RightBeforeWithering
  - LetTheBerryDie

The latest is now used for the Kasib unlock