Hatchery: Fixup pokerus spreading feature
The partyPokemon member access was changed to a KnockoutObservable by
pokeclicker/pokeclicker@ad2555ee764c3ec4c8376c7a00adef8fa60c0cec

This resulted in not detecting currently breeding infected pokemons,
which prevented corresponding pokemons to be added.
For some reason this was not caught by the tests,
the behavior might be browser-dependent.

The member is now accessed properly