Mine: Use enough restore item to use a bomb
In early game, the small restore only gives 5 energy.
A bomb costs 10 energy, and if the player has not enough energy to use
one, the automation is stopped until the next loop, up to 10s later.

Many restores can now be used in the same loop to make sure that the
player always has enough energy, if he can.