Mine: Force the daily deals to run immediatly
The daily deals interval is set to 5minutes.
The first run did happen 5min after enabling the feature.
If a setting was changed, it was only visible by the user at the next
run, which could happen 5min later.

The feature is now run immediately after activating an option.
Additionally, enabling a setting will trigger it again.