Dungeon: Improve dungeon speed with flashlight
When the player ahs the flashlight unlocked (after 200 clears), he can
see every adjacent tiles to the ones he visited.
Instead of simply visiting every tiles of the dungeon, the automation
will choose the most suited tile to visite next.
Such tile is the one that will reveal the most hidden tiles if moved to