Mine: Improve the hammer use efficiency
The hammer was not be considered for revealed tiles.
This could lead in hammer use discard, even though it was the best
move.

Lets take the following setup example:
    111
    101
    111

With the previous logic this would use the equivalent of 2 hammer
actions worth of energy using the chisel instead of the hammer.

The hammer is now considered on each tile, resulting in a single
hammer use in the above-mentionned example.