Trivia: Add a button to move to the roamers route
The roamers route name can be pretty long in the later regions.
The name could overflow the menu width.
Additionally, the route names are not only numbers, and finding
the route location can be a bit tricky (even more with subregions).

The route name in now displayed in the tooltip and a button is
available to jump right there.