Focus.Achievements: Fixup achievement never being selected
The achievement region was not retrieved using the right path.
The value would then be undefined.

If the given region was undefined, the canMoveToRegion would return
true on some browser.

The canMoveToRegion function now make sure the given region is a number
to ensure a consistent behaviour between browsers.

The region property is now accessed correctly.