Replace use of instanceof keyword
The next pokeclicker version will use module instead of raw javascript.
Those modules will only be accessible by the script that imported them.

This prevents the use of instanceof keyword to work,
since the class needs to be known to determine its instance this way.

A new utils method isInstanceOf was added to use the constructor name
of the class, which doesn't require to know the class but only its name

This method iterate over the object internal __proto__ member to
progress through inheritance.