Focus.Quests: Move to the appropriate place to farm money
When the player does not have balls anymore and needs to buy some
but does not have enough money to do so, it was still moved to the
pokemon catching place, with the corresponding loadout.

It will now move to the money farming place and equip the money
farming loadout.