Focus.Quest: fixup farming automation always being active
If the user disabled every farming quests, the farming automation
should not be enabled.
This automation did not take over the farming behaviour by forcing
the replant mode, but did enable the farming automation, even though
the user disabled it.

The automation state is now respected when no farming quests are
selected.