Focus.PokerusCure: Add the Focus on pokérus feature
This feature will automatically move to routes were there is at least
one infected pokemon that is not resistant to the pokérus.
It will equipe the set pokemon catching pokéball and catch every
pokémon on that route until the full route is resistant.
It will then move to the next route.