Dungeon: Restore the flash-light strategy
Since v0.10.9 the flashlight can reveal tiles beyond the immediate
nearby tile. This caused the player not being able to move to the
most beneficial tile in most case, since it would not be accessible.

Only the accessible tiles are now considered instead of all visible
tiles.

This fixup still has some flows with the chest collection process.