Dungeon: Don't stop anymore on user action when having the flash light
To avoid breaking the regular dungeon strategy, user action had to be
detected to ensure proper dungeon completion.

In case the user has the flash-light with at least the second update
(introduced in v0.10.9), the user action detection can be triggered
when the player did not do anything.
This check is not relevant and have been removed in such case.