Hatchery: Don't farm unlocked mega-evolution in real evol challenge
In real evolution challenge, the base pokémon stats gets transferred
to the evolution. The base attack of the base pokémon gets reset.

The hatchery will no longer consider those pokémons if the challenge
is enabled.

In the standard mode, the already unlocked evolutions will still be
selected. This is done because the game might change the requirements
and this would keep players from being able to farm the shiny form
this way.