Farm: Display the plantation strategy in the floating panel
The user might wander what strategy is currently used by the automation
to unlock the currently selected berry.

This information is now displayed under the current berry name in the
floating panel, for the mutation strategies.

Hovering a berry will display a tooltip with its name and the delay
before planting it compared to the first planted berry (if applicable)