Focus.Quest: Add support for the new BattleFrontier quest
v0.10.14 introduced ClearBattleFrontier quests.

When this quest is active, it will be picked after any other
handled type of quests.
The player will be moved to Battle Frontier town.
The BattleFrontier mode will be entered and the session will start
automatically (from the last checkpoint, if any).

When the quest objective is met, or the automation is turned off,
the BattleFrontier mode will be left.