Safari: Fixup problem when entering a new instance
When an instance of the Safari was completed, and a new one
was started, the path to the farm area was not refreshed.
This resulted in the player going to the previous run location,
often not a Grass of Water location.

This could cause crashes as well, if the user resized their browser
window, since the size of the safari grid depends on the browser size.

The internal data is now properly reset between each run.