Mine: Remove daily deals automation
The v0.10.23 update includes a major rework to the Underground feature.
That includes completely removing daily deals and chains.
Instead, players will be able to trade items of the same type
at a rate of 3 to 1 at any time.

For example, shards can be traded for other shards, plates can either
be sold for gems or traded for other plates,
and diamond items can only be sold for diamonds.