Dungeon: Don't change catch settings during automation mode
If the user sets a pokéball type to catch bosses and the automation
forced a dungeon run, it would cause the automation setting to be
overridden and it would never go back to the proper one.

The catch settings are now only changed if no automation mode is set.